Sample feedback (#17)
* Remove Unused files

* Update Readme.md

* Update Gemfile

* Update Application Branding

* Use secrets instead of ENV

* Apply feedback on Custom Login

* Update secured controller by secured concern

* Update documentation

* Update 02 - Custom Login - Remove all OmniAuth dependencies

* Updates ready files, fixes comments in methods.
