+++ 01-Login/README.md
@@ -8,16 +8,17 @@ In order to run the example you need to have ruby installed.
 
 You also need to set the ClientSecret, ClientId, Domain and CallbackURL for your Auth0 app as environment variables with the following names respectively: `AUTH0_CLIENT_SECRET`, `AUTH0_CLIENT_ID`, `AUTH0_DOMAIN` and `AUTH0_CALLBACK_URL`.
 
-For that, if you just create a file named `.env` in the project directory and set the values as follows, the app will just work:
+Set the environment variables in `.env` to match those your Auth0 Client.
 
 ````bash
 # .env file
-AUTH0_CLIENT_SECRET=myCoolSecret
 AUTH0_CLIENT_ID=myCoolClientId
+AUTH0_CLIENT_SECRET=myCoolSecret
 AUTH0_DOMAIN=samples.auth0.com
 AUTH0_CALLBACK_URL=http://localhost:3000/auth/auth0/callback
 ````
 Once you've set those 4 environment variables, run `bundle install` and then `rails s`. Now, browse [http://localhost:3000/](http://localhost:3000/).
+__Note:__ Remember that you need to have `./bin` in your path for `rails s` to work.
 
 Shut it down manually with Ctrl-C.
 
@@ -26,6 +27,7 @@ __Note:__ If you are using Windows, uncomment the `tzinfo-data` gem in the gemfi
 ## Important Snippets
 
 ### 1. Auth0 Lock Setup
+[Home Javascript Code](/01-Login/app/assets/javascripts/home.js.erb)
 ```js
 var options = {
    auth: {
@@ -41,14 +43,15 @@ function signin() {
  	lock.show();
 }
 ```
-### 2. Check if  User is Authenticated in Secured Controller
-
+### 2. Check if  User is Authenticated in Secured Controller Concern
+[Secured Controller Concern Code](/01-Login/app/controllers/concerns/secured.rb)
 ```ruby
-#Note: Controllers with secured actions will need to inherit from Secured Controller.
-class SecuredController < ApplicationController
-  before_action :logged_in_using_omniauth?
+module Secured
+  extend ActiveSupport::Concern
 
-  private
+  included do
+    before_action :logged_in_using_omniauth?
+  end
 
   def logged_in_using_omniauth?
     redirect_to '/' unless session[:userinfo].present?
@@ -57,17 +60,22 @@ end
 ```
 
 ### 3. Store User Profile Upon Successful Authentication
+[Auth0 Controller Code](/01-Login/app/controllers/auth0_controller.rb)
 ```ruby
-class Auth0Controller < ApplicationController
-  def callback
-    session[:userinfo] = request.env['omniauth.auth']
+def callback
+  # OmniAuth places the User Profile information (retrieved by omniauth-auth0) in request.env['omniauth.auth'].
+  # In this tutorial, you will store that info in the session, under 'userinfo'.
+  # If the id_token is needed, you can get it from session[:userinfo]['credentials']['id_token'].
+  # Refer to https://github.com/auth0/omniauth-auth0#auth-hash for complete information on 'omniauth.auth' contents.
+  session[:userinfo] = request.env['omniauth.auth']
 
-    redirect_to '/dashboard'
-  end
+  redirect_to '/dashboard'
+end
 
-  def failure
-    @error_msg = request.params['message']
-  end
+# if user authentication fails on the provider side OmniAuth will redirect to /auth/failure,
+# passing the error message in the 'message' request param.
+def failure
+  @error_msg = request.params['message']
 end
 ```
