+++ 01-Login/app/controllers/dashboard_controller.rb
@@ -1,4 +1,6 @@
-class DashboardController < SecuredController
+class DashboardController < ApplicationController
+ include Secured
+
   def show
   end
 end
