+++ 02-Custom-Login/app/assets/javascripts/application.js
@@ -10,4 +10,4 @@
 // Read Sprockets README (https://github.com/sstephenson/sprockets#sprockets-directives) for details
 // about supported directives.
 //
-//= require_tree .
+//= require_tree
