+++ 02-Custom-Login/app/controllers/auth0_controller.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 class Auth0Controller < ApplicationController
+  # Once the login or login with Google is clicked this method is called.
   def callback
     if params[:user]
       signup if params[:signup]
@@ -14,24 +15,25 @@ class Auth0Controller < ApplicationController
     redirect_to '/', notice: ex.message
   end
 
+  # Used to redirect to auth0 authorization url to login with Google.
   def google_authorize
     redirect_to client.authorization_url(
-      ENV['AUTH0_CALLBACK_URL'],
+      Rails.application.secrets.auth0_callback_url,
       connection: 'google-oauth2',
       scope: 'openid'
     ).to_s
   end
 
-  def failure
-    @error_msg = request.params['message']
-  end
-
   private
 
+  # Gets the user token when logging in using google-oauth2.
+  # @see https://auth0.com/docs/auth-api#!#post--oauth-access_token
+  # @see https://github.com/auth0/ruby-auth0/blob/master/lib/auth0/api/authentication_endpoints.rb
   def google_login
-    client.obtain_user_tokens(params['code'], ENV['AUTH0_CALLBACK_URL'], 'google-oauth2', 'openid')['id_token']
+    client.obtain_user_tokens(params['code'], Rails.application.secrets.auth0_callback_url, 'google-oauth2', 'openid')['id_token']
   end
 
+  # Login with username / password using the Auth0-Ruby SDK.
   def login
     client.login(
       params[:user],
@@ -43,6 +45,9 @@ class Auth0Controller < ApplicationController
     )
   end
 
+  # Signs a new user up using the Auth0-Ruby SDK (username / password).
+  # @see https://auth0.com/docs/auth-api#!#post--dbconnections-signup
+  # @see https://github.com/auth0/ruby-auth0/blob/master/lib/auth0/api/authentication_endpoints.rb
   def signup
     client.signup(
       params[:user],
