+++ 02-Custom-Login/app/controllers/concerns/secured.rb
@@ -1,0 +1,11 @@
+module Secured
+  extend ActiveSupport::Concern
+
+  included do
+    before_action :logged_in?
+  end
+
+  def logged_in?
+    redirect_to '/' unless session[:token_id].present?
+  end
+end
