+++ 02-Custom-Login/app/controllers/dashboard_controller.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
-class DashboardController < SecuredController
+class DashboardController < ApplicationController
+  include Secured
   def show
   end
 end
