--- 02-Custom-Login/app/controllers/secured_controller.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-class SecuredController < ApplicationController
-  before_action :logged_in?
-
-  private
-
-  def logged_in?
-    redirect_to '/' unless session[:token_id].present?
-  end
-end
