--- 02-Custom-Login/app/views/auth0/failure.html.erb
@@ -1,3 +1,0 @@
-<div class="panel panel-danger">
-  <h2>Error <%= @error_msg %></h2>
-</div>
