+++ 02-Custom-Login/app/views/layouts/application.html.erb
@@ -6,6 +6,8 @@
   <%= yield :cdn_code %>
   <%= stylesheet_link_tag    '//maxcdn.bootstrapcdn.com/bootstrap/3.2.0/css/bootstrap.min.css' %>
   <%= stylesheet_link_tag    '//maxcdn.bootstrapcdn.com/font-awesome/4.1.0/css/font-awesome.min.css' %>
+  <%= javascript_include_tag '//ajax.googleapis.com/ajax/libs/jquery/1.12.4/jquery.min.js' %>
+  <%= javascript_include_tag '//maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js' %>  
   <%= javascript_include_tag '//cdn.auth0.com/js/lock/10.0.0/lock.min.js' %>
   <%= javascript_include_tag 'application' %>
   <%= csrf_meta_tags %>
@@ -20,7 +22,7 @@
             <span class="icon-bar"></span>
             <span class="icon-bar"></span>
           </button>
-          <a class="navbar-brand" href="http://auth0.com">Auth0</a>
+          <a class="navbar-brand" href="#">My Application</a>
         </div>
         <div id="navbar" class="navbar-collapse collapse">
           <ul class="nav navbar-nav">
