+++ 02-Custom-Login/bin/rake
@@ -1,4 +1,5 @@
 #!/usr/bin/env ruby
+# frozen_string_literal: true
 require_relative '../config/boot'
 require 'rake'
 Rake.application.run
