+++ 04-User-Profile/app/controllers/concerns/secured.rb
@@ -1,0 +1,11 @@
+module Secured
+  extend ActiveSupport::Concern
+
+  included do
+    before_action :logged_in_using_omniauth?
+  end
+
+  def logged_in_using_omniauth?
+    redirect_to '/' unless session[:userinfo].present?
+  end
+end
