+++ 04-User-Profile/app/controllers/dashboard_controller.rb
@@ -1,4 +1,5 @@
-class DashboardController < SecuredController
+class DashboardController < ApplicationController
+  include Secured
   def show
     @user = session[:userinfo]
   end
