+++ 04-User-Profile/app/views/layouts/application.html.erb
@@ -24,7 +24,7 @@
             <span class="icon-bar"></span>
             <span class="icon-bar"></span>
           </button>
-          <a class="navbar-brand" href="http://auth0.com">Auth0</a>
+          <a class="navbar-brand" href="#">My Application</a>
         </div>
         <div id="navbar" class="navbar-collapse collapse">
           <ul class="nav navbar-nav">
